+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Refactor saved state logic to its own module
+- Refactor react styles to its own module
+
 ### Fixed
 
 - Stop fetching degrees twice